requirements.txt
flask
python-dotenv
flask_sqlalchemy
psycopg2-binary
pytest
pytest-cov